Move tlBank UI to client-only component
Extract the large TL Bank UI into components/TlBankPage.tsx and load it client-side via dynamic import in pages/tlBank.tsx (ssr: false) to avoid running web3-onboard hooks during SSR. Add a local Next API at pages/api/news.ts to serve a mock article for dev/tests, and improve production CF Pages function (functions/api/news.ts) to parse stored KV payloads defensively. Normalize news response handling in pages/news.tsx. Remove the client-side redirect logic from pages/_app.tsx that forced navigation to /tlBank. Also update FE Robot tests (listen_page.robot, news_page.robot), TEST/devplan.md, and .gitignore to include new test result paths.
