+++ TEST/tests/fe/listen_page.robot
@@ -1,17 +1,19 @@
 *** Settings ***
-Library           Browser
-Resource          ../../resources/variables.robot
-Suite Setup       New Browser    chromium    headless=True
-Suite Teardown    Close Browser
+Documentation    Listen page — Bankless Africa copy appears in page body (no fragile scroll).
+Library    Browser    timeout=60s
+Resource    ../../resources/variables.robot
+
+Suite Setup       Open Listen Browser
+Suite Teardown    Close Browser    ALL
+
+*** Keywords ***
+Open Listen Browser
+    New Browser    headless=true    args=["--no-sandbox", "--disable-dev-shm-usage"]
 
 *** Test Cases ***
 Listen Page Should Display Bankless Africa Podcast
     New Page    ${BASE_URL}/listen
-    Sleep    5s
-    Get Title    should be    BanklessDAO Community
-    ${bankless_africa_section}=    Get Element    text=Bankless Africa
-    Should Be True    ${bankless_africa_section} is not None
-    Within    ${bankless_africa_section}
-        Get Text    text=Hosted By
-        Get Text    text=Exploring the world of Web3 and DeFi in Africa.
-    END
+    Wait For Load State    networkidle
+    ${body}=    Get Text    body
+    Should Contain    ${body}    Bankless Africa
+    Should Contain    ${body}    Exploring the world of Web3 and DeFi in Africa.
